UPSTREAM: wireguard: queueing: do not account for pfmemalloc when clearing skb header
Before 8b7008620b84 ("net: Don't copy pfmemalloc flag in __copy_skb_
header()"), the pfmemalloc flag used to be between headers_start and
headers_end, which is a region we clear when preparing the packet for
encryption/decryption. This is a parameter we certainly want to
preserve, which is why 8b7008620b84 moved it out of there. The code here
was written in a world before 8b7008620b84, though, where we had to
manually account for it. This commit brings things up to speed.

(cherry picked from commit 04d2ea92a18417619182cbb79063f154892b0150)
Bug: 152722841
Change-Id: I1f2c3176c00b01adae90143d958f1d98c7fea8c9
Signed-off-by: Kunmun <kunmun.devroms@gmail.com>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: meydiwahendra <meydiwahendra@gmail.com>