arm64: Inline the spinlock function family regardless of preemption
According to Sultan Alsawaf, x86 guarded spinlock inlinig behind !PREEMPT
back in 2005 due to possible instability and debugging requirements [1],
however this no longer applies. Inling spinlocks provides a performance
improvement without reducing the kernel's stability, even when preemption
is enabled.

Futhermore, while it's true that preemption increases the size of the
spinlock code, (which in turn causes it to be less likely to hit the
cache) inlining the spinlock function family still improves performance
according to Sultan [2] with a "5% boost to procfs I/O performance".

[1] : https://lore.kernel.org/lkml/20200530050943.142128-1-sultan@kerneltoast.com
[2] : https://github.com/kerneltoast/android_kernel_google_wahoo/commit/2de3e049b6858c3c56582b7a73b7b9b50c9a5dc8

Signed-off-by: Diab Neiroukh <lazerl0rd@thezest.dev>
Signed-off-by: Carlos Ayrton Lopez Arroyo <15030201@itcelaya.edu.mx>
Signed-off-by: DennySPb <dennyspb@gmail.com>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: meydiwahendra <meydiwahendra@gmail.com>