techpack: audio: Reduce verbosity of logging
These are just useless. There are some logs that show up in dmesg
meaning that there are some bugs introduced by CAF's techpack.
At the end, everything works. Demote the useless logs to debug including
the ones that let us know about the bug until we find a fix to them, as
they are quite spammy.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: meydiwahendra <meydiwahendra@gmail.com>