vidc_3x: Instantiate tracepoints separately when CONFIG_TRACEPOINTS=y
Tracepoints was defined in msm_vidc_debug which had been compiled out
using CONFIG_DEBUG_KERNEL macro. So, if we replace the macro with
CONFIG_TRACEPOINTS, then it will resolve the compilation error when
Ftrace is enabled but will compile the other debugs alongside.
Therefore, create a dedicated msm_vidc_trace.c for tracepoints
and compile it when CONFIG_TRACEPOINTS is enabled.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: meydiwahendra <meydiwahendra@gmail.com>